Improve /junction functionality
This adds a /junction list command to show the lines for the current
junction.

This causes right clicking on a rail block associated with a junction
sign to select that junction in the junction editor.

This functionality will help rail managers who want to edit hidden
junction signs without digging.